Fix cluster-health-verification with poetry2nix
The cluster-health-verification triggers actions scripts which in turn
start Ansible playbooks. Somewhere during this cascade, the variable
ansible_playbook_interpreter gets lost so the inner Ansible uses the
wrong interpreter.

This commit introduces an enviroment variable ANSIBLE_PYTHON_INTERPRETER
for the verify-cluster-health.sh action script which unfortunately
Ansible itself does not provide[1]. Using this variable, it is possible
to pass down the correct interpreter to the inner Ansible.

[1] https://github.com/ansible/ansible/issues/6345#issuecomment-37129460