Fix devshell reloading when updating python deps
When modifying Python dependencies, the direnv devShell is expected to
update automatically. Python specific files have been watched only
inside layout_poetry but not inside use_flake_if_nix.

This commit adds the respective watch_file invocations to the latter
using a function to deduplicate between both layouts.

In this process, the environment variable PYPROJECT_TOML which was never
documented anywhere and also pretty useless was removed as keeping it
around would be prone to bugs.