ci: push to nix cache in background
Currently, we wait until everything has been built and the container
image has been pushed to the registry before we push the build artifacts
to the Nix binary cache. Hence, in case of failure throughout the run,
everything needs to be rebuilt.

This commit changes the behavior such that new build artifacts are
pushed to the Nix cache in the background as soon as they are
successfully built. Consecutive runs will now fetch them instead of
rebuilding. Repeating failed runs is thus sped up.