Move Vault S3 backup configuration to Vault
The Vault S3 backup configuration, containing sensitive information,
has been moved out of the cluster repository to Vault for security.

A migration function is provided to assist with moving existing
configurations.

The documentation has been updated with relevant details.

This change is classified as a breaking change because existing
clusters will require an update to their Vault policies.

Closes: #584