Fix K8s installation fact for existing clusters
We switched to static facts in !1450. However, for existing clusters
workers will try to join Kubernetes on re-runs as there is no static
fact, yet. We therefore have to add migration tasks to set the static
fact, if a node already joined Kubernetes. For control-plane nodes this
is not needed as we check if the remote or local kube-apiserver is
reachable and skip the join process based on that info.

We check the kubelet service as this should be a sufficient indicator.
We must not check the existens of a proper kubelet config, because that
gets generated by Vault nowadays and is always present, even if the node
did not join already.