Fix k8s-worker join for existing clusters
For clusters setup before release v8.1, the static fact that it already
joined the Kubernetes cluster is not available for a worker. We created
a hotfix for that, but that was insufficient. Before a node actually
joined the Kubernetes cluster, kubelet is already enabled and is
starting and stopping. So checking for that is not sufficient to know
if a node is already part of the cluster.