Fix system update of new nodes
As we moved to static facts and changed code structure, we can't simlpy
intersect the groups here anymore. We must group for both initialization
facts as otherwise the intersection may be an empty set when
initializing the cluster (the facts are not set in that case and the
groups are empty). As we still want to prevent to update Kubernetes
nodes which act as frontend without draining them, we must do a proper
intersection of both facts.