Fix fetching root CA key if signing requests are enabled
The helpers/vault-approle is executed in advance for renewal of the
Kubernetes PKI, but there was a mismatch in the variable names. As
everywhere else we set "vault_role_id" to "vault_node_role_id" (which is
set by the helper role), we do the same here as well.
This did not affect an rollout of k8s-core, but only Kubernetes
upgrades.