Add CI job to render options reference
Since !1265, the option reference for the documentation is now rendered
via Nix. Because they could be reproduces on-the-fly, the resulting RST
files were not checked into git. However, implementing the rendering
step with sphinx-multiversion has proved quite hard, thus it has been
decided to check the rendered reference into git for now.

This commits adds a new CI stage "housekeeping" containing a new job to
render and commit the options reference if it has changed.