Drop poetry
After dropping poetry2nix and going back to the pure poetry direnv
layout, we are back to its drawbacks including impure builds, the
occasional hard-to-debug "it works on my machine" build error, problems
with the keychain and most importantly lengthy package installations
with non-ideal deduplication.

This commit implements option 4 as discussed in [1]. It drops all poetry
usage from the repository and adds all Python dependencies via nixpkgs
directly.

One exception is kubernetes-validate which has been submitted to
nixpkgs[1]. Until it's accepted, its package definition is added to
this repository.

[1] https://gitlab.com/yaook/k8s/-/issues/723
[2] https://github.com/NixOS/nixpkgs/pull/356174