Rework requirements + devShells
When the flake.nix was introduced in this repository, its only purpose
was the then-optional devShell with a couple of dependencies. We have
since
* completed the dependency listing to be exhaustive
* introduced multiple dependency groups (minimal, ci, interactive)
* started to build the CI image using Nix
* migrated to a Nix based configuration.

All of this has made the flake.nix file quite lengthy over time, which
should be a concise and overseable entry-point. Also, the different
groups are each treated specially, which ideally should not be the case.

This commit introduces a new Flake module yk8s-env which allows to
define dependecy groups as a DAG. Each dependency group will produce
it's own output package and devShell to be consumed by the direnv layout
or other Nix expressions.