(tmp) try without direnv in CI
This is a temporary commit to be either dropped or squashed onto Drop
poetry.

Now that all Python dependencies are defined in Nix, the CI image
contains everything that is needed to run YAOOK/K8s. This means that we
can drop the direnv invocations, which will speed up CI runs because
they won't have to evaluate Nix expressions nor download Python
packages.