Bump Thanos
Since [1] PDBs are enabled by default which is inconvenient because the
default replicas are also 1 so the default values basically block the
drain of a node. As we do not specify an option to scale up Thanos
components and had no issues with that so far, we also disable PDBs.

https://github.com/bitnami/charts/blob/main/bitnami/thanos/CHANGELOG.md#1570-2024-06-05