Adjust Kubernetes upgrade procedure
We now more strictly follow the official kubeadm upgrade guide.
The previous approach resulted in a partially updated control plane
as all but the first control plane node still ran components of the
previous Kubernetes version. This is because:
When "kubeadm upgrade apply" is triggered, the ClusterConfiguration is
updated. "kubeadm upgrade node" uses the ClusterConfiguration to
determine the Kubernetes version. Therefore, "kubeadm upgrade apply"
must be run prior to "kubeadm upgrade node".
However, for "kubeadm upgrade apply" to succeed, we must update
kubeadm (the package) on all control plane nodes before triggering it
as otherwise the upgrade fails as the mode to upgrade addons has been
changed since v1.28 and forces us to update the kubeadm package in advance.
Only after the ClusterConfiguration is updated for the next_k8s_version,
"kubeadm upgrade node" will update the control plane components to the
correct version.

For "kubeadm upgrade" in general it is not necessary to drain the node.
Therefore, the procedure has been adjusted such that the kubeadm package
is updated and the kubeadm upgrade commands are triggered without a drain.

https://kubernetes.io/docs/tasks/administer-cluster/kubeadm/kubeadm-upgrade/