Create script to create CSRs for intermediate CAs
For clusters using intermediate CAs, the CAs are valid for 1.5 years.
As the root CA usually has a longer lifespan, they must be updated
frequently. This script can be used to create new CSRs which then can be
signed by the externally managed root CA and afterwards loaded back to
vault. This script is basically a wrapper for the "mkcsrs" function of
"lib.sh".