Fix import of signed intermediates
In 8153896, support for rotating the root CA of clusters with
intermediate certificates was added. There, the signature of the
`import_cert` function was added to include the issuer name.

Unfortunately, the `load-signed-intermediates.sh` script was not updated
for that.

After reading the code, I'm reasonably confident that the correct issuer
name for a freshly created cluster with signed intermediates is
`current` (based on what the root CA rotation assigns for the default
issuer).