Test direnv setup in CI
Currently, the container image is built with all dependencies in PATH
which means that inside of it, YAOOK/K8s can be used out-of-the-box.

This may be desirable for images targeted at interactive use, but for
the CI it means that the direnv setup cannot properly tested.

This commit changes the container image so that is is built with all
dependencies available in the Nix store but not in PATH.