ci: Don't run K8s upgrade jobs on MRs
Adjusting the rules for the K8s upgrade jobs such that they are not
obligatory for merge requests/merge trains anymore. This is due to the
fact that these jobs are currently error-prone and consume a lot of time
which slows down overall development pace. Instead, we configure a
scheduled job against devel such that they are executed once a week to
still be able to detect regressions. The jobs are still mandatory for
release pipelines though.