Don't update inventory for vault-in-docker scripts
In c0be6c50d, all action scripts have been added a step to update the
inventory. As a result, in devclusters using vault-in-docker, each
reload of the environment caused an evaluation of the inventory which
slowed down't the reload significantly. Both scripts vault.sh and
vault_env.sh, however don't need access to the inventory.

This commit removes the call to update-inventory.sh from both vault.sh
and vault_env.sh.