ci: ensure correct image is used
Previously, the build_image job was only triggered if changes were
detected in its source files. Otherwise, the "devel" tag was used as a
default. This behavior caused problems in hotfix pipelines of release
branches that were no longer compatible with the CI image on devel, see

Starting with this commit, the build_image job is always run as the
first job in a pipeline. It evaluates the ciImage derivation and uses
its hash to check if the image already exists in the registry. Only if
it does not, it is built and pushed.

This also means that images are no longer tagged with their commit hash
but with the hash of the output derivation, so there is only one tag per
digest.

Closes: #745
Cherry-picked-from: 0e93e1a6 ci: ensure correct image is used
Includes: f8571988 ci: fix pages job
Includes: 0c0adda7 ci: don't tag container images with commit ref names