CI: Use csi-sc-cinderplugin for Grafana
We frequently see the monitoring tests fail to scrape the Grafana
endpoint. rook-ceph-cephfs in clusters running on top of OpenStack is
known to have bad performance (because it is basically Ceph-on-Ceph).
I assume that cephfs may not be performant enough for Grafana which is
why the tests fail. Therefore, this commit changes the storage class
used for Grafana. CephFS itself is tested in a dedicated testing role.