ci: fix pages job
Starting with !1627, every pipeline starts with the build_image job
which determines the image tag and ensures it is present in the
repository. The image tag is propagated to the other jobs via a build
artifact. The pages job was configured to not fetch any artifacts from
previous jobs in order to ensure a clean public/ directory.

This commit adds build_image as a dependency for pages, so that the
image tag is present for the job.