ci: don't tag container images with commit ref names
Starting with !1627, each pipeline uses a reproducible image tag. As a
convenience for developers, images were additionally tagged with the
commit ref name (eg. tag or branch name). However, this lead to
difficulties because ref names can contain characters that are not
allowed in OCI image tag names.

This commit removes the command that tags images with the ref name as it
is not needed for the CI and would create too much hassle for a
nice-to-have feature.