Use rolling python3 version
Previously, we pinned our Python version to 3.11 which resulted in long
build times for everyone (including the CI) when 3.11 packages vanished
from the nixos cache.

This commit unpins Python 3.11 and sets the package to "python3" which
resolves to the Python version currently targeted by nixpkgs. This way,
we stay up-to-date with the current version and avoid having to build
old packages from source.