Revert "Add minimal env to ci env"
This reverts commit dd091c106cfd18d3d57be2e5b4f1df84ca878cd3.

In 6ec0bc14bf2d3968704eac933c52a2932ceb5f0e (as part of !1570), we
decided to add all dependencies to the image but not to the PATH in
order to test the direnv setup. That's why the ci group only includes
docs and lint, but not any of the other groups. Any additional
dependencies needed by the CI have to be listed individually.