ci: replace cleanup script with upstream command
Historically, cleaning up an OpenStack project in the correct order was
challenging due to numerous dependencies between objects. To address
this, we developed a script to handle the clean-up process.

However, the openstack-client now provides a subcommand with equivalent
functionality. This commit replaces our custom script with the official
subcommand.