Fix wireguard module for non-wireguard clusters
The wireguard module contained a bug affecting clusters with no
wireguard configuration. Even if disabled, the module still tried to
build the inventory and state file using the wireguard helper which in
turn failed due to no peers being configured.

This commit adds conditionals for both inventory and state files so that
the wireguard helper is only invoked if wireguard is enabled. In case it
is disabled, an inventory file containing {enabled = false;} is
generated so load_conf_vars keeps working.