CI: Don't publish pages in MRs
Introducing a dry-run of the pages job which does not publish the
"public/" directory as artifact which would tinker with Gitlab pages
[1].
For code deduplication, a newly introduced ".pages" template is used for
the dry-run as well as the actual pages job.
The dry-run is executed in stage "lint" because that's probably the best
fitting stage.

[1] https://docs.gitlab.com/ee/user/project/pages/#how-it-works