migrations: Fix Terraform transition for remote backend
The migration script in its previous version was not able to handle
clusters which have Gitlab configured as Terraform backend.
Tinkering with the Terraform state is only doable with a local state
file. Therefore, we have to temporarily migrate from remote to local
backend. To reduce code duplication, some of the functions have been
moved from "apply-terraform.sh" to "lib.sh".

Closes: #749