Fix tf_usage detection for bare-metal
!1265 introduced a new inventory file  at
yaook-k8s/group_vars/all/terraform.yaml that contains the enabled state
of terraform. This file should be present in any case so it can be
reliably detected whether terraform is enabled or not. Due to a bug,
this file was only generated if terraform was enabled.

This commit removes the erroneous condition, so the file is always
generated.