Create helper role for kubeadm commands
At several points during life cycling Kubernetes we have to trigger kubeadm
commands. This commit introduces a helper role which can be used as a
wrapper. The helper role does some crucial verifications and in the case
of control plane nodes, it also ensures the ca.key is properly handled
before and after executing a kubeadm command. It does also ensure a
umask is set. This makes executing kubeadm more stringent and less
error-prone. Furthermore, this also reduces code deduplication.