Stop filtering null values during inventory generation
In order to stay as close as possible to the logic from
update_inventory.py, I added a filterNull transformation that removes
values with null value from the inventory. This means that conditions in
Ansible need to check for is defined.
Unfortunately this makes it easy to introduce bugs that are easy to
overlook during review.

Thus, this commit ensures that all variables are defined within the
inventory with explicit null values instead of leaving them out.

Closes #756