Fix testing.nodes
For a long time, we've been providing an option to limit the test stage
to certain nodes. Historically, this was done in the dictionary
`testing.test-nodes`. This option was replaced by the list
`testing.nodes`. However, in config.template.toml we continued to call
it `testing.test-nodes`. Thus, the feature was non-functional if
configured as documented and when transforming the config to Nix, the
mistake was inherited.

This commit renames the option to `testing.nodes`.