Improve advice about group membership during nix installation
Just restarting the terminal session is not sufficient. Terminals are
generally forked off the window manager's, desktop shell's or
compositor's process, which has gained its identity during initial
login. Group memberships are not updated on running processes, so there
is no way those could've gained the new group ID.

The only way to obtain the group ID on the entire desktop session is to
restart it.