Move finding out of orchestrator role to helper role
For the "k8s-login" role (triggered in "k8s-master" role) to succeed,
it must be known if the token of the orchestrator has the orchestrator
role. Therefore, the finding out is moved to a helper role. This fixes a
bug which caused the "install-k8s" playbook to fail if explicitly
triggered.

Rel: #765