flake: Use nixos-24.11
We decided to switch to NixOS stable. This has already been done on
devel. However, we decided to backport that.
One reason for that is that we're using boto3 to manage the S3 bucket
for etcd-backups, but the latest versions of boto3 are incompatible
to OpenStack Swift. To buy us some time to circumenvent that, we're
switching to NixOS stable which will give as an older version of boto3
that is still compatible.