Use canReach instead of cidrs to find node IP addresses
Using cidrs works in many cases, but it breaks down when VRRP is in use
on the nodes. This happens in bare metal setups where keepalived is
installed on the control plane.

The keepalived/VRRP address is typically from the same subnet as the
node addresses (it has to be, even). Thus, the CIDR will generally also
match the VRRP address. This is not necessarily critical: only one of
the nodes has it at any given point in time and thus everything works
out.

However, if the VRRP-primary node is down for a long time, calico-node
may pick up the VRRP address on another node. Then, when the
VRRP-primary returns, it fails to boot its calico-node because calico
detects the VRRP address as node address which is already assigned to
another node.

Hence, we have to use a different method. We opt to use the
[can-reach](https://docs.tigera.io/calico/latest/reference/configure-calico-node#can-reachdestination)
method here, and we fill in the network address of the subnet CIDR.

Now you may rightfully ask: Why does that even work? Surely, and in
particular with IPv4, the network address (the 0th address in a range)
is not reachable, so can-reach would never find anything. Luckily,
calico does not *actually* try to reach that address. Instead, it uses
the local routing table.

Any address assigned to an interface creates a route for its prefix.
When you assign 2001:db8::abcd/112, a route for 2001:db8::/112 is
created on that interface automatically, using 2001:db8::abcd as source
address. However, if you assign 2001:db8::def0/128, no such route is
created, because the prefix has only that one address.

VRRP addresses use single-address-prefixes (/128 in IPv6 and /32 in
IPv4), so they won't ever be found by can-reach. can-reach will only
ever find the "real" address which has a prefix with more than one
address.