Decouple helm invocations from local state
Currently, the behavior of deploying Helm charts is inconsistent across
different roles. Some directly reference the upstream URL of the
repository, some add the repository to the local Helm state first and
then refer to that and some add it to the local state but never use it
and refer to the upstream URL anyway.

Using the local Helm state has a few downsides:
* If the local Helm state is broken, the run will fail
* If a repository with the same name has already been configured
locally, but with a different URL, this URL will be overwritten
* If `update_repo_cache: true` is forgotten, a stale version of the
local copy of the repo will be used

This commit unifies the behavior and decouples all helm invocations
completely from local state by referring directly to the upstream URL.