Add support for optional basic auth secret for remote writes
We want to have authenticated remote-write endpoints. Basic auth is the
simplest way to do so.

Note that we are already using basic auth in some places, but we are
working around the LCM restriction of not being able to set it by
spinning up a second prometheus. That is wasteful.