Split inventory generation into multiple targets
This commit splits inventory generation into multiple targets for eg.
vault, terraform and ansible. This allows for better isolation and also
for better dependency checking. For example previously, non-existing
Terraform outputs had to be ignored because Terraform had to be run
first for them to be generated. Now, the Terraform target can be used to
provide the variables to Terraform and then the Ansible target
_requires_ the Terraform outputs.