Set static Ansible fact after initializing K8s
In 9fce13d0 we introduced the usage of Ansible static facts to
persistently save information about the Kubernetes initialization status
of a node.

However, it went unnoticed until now that the static fact is not set on
the first CP node after initializing Kubernetes. This commit fixes that.

This went unnoticed because until !1788 we had migration tasks which
dynamically checked if Kubernetes already has been initialized. No harm
done therefore.