Fix Terraform module for bare-metal after refactor
efa470766140408504986656e842e1b444471de9 moved the generation of
Terraform state files and the Terraform based inventory from the
OpenStack module to the Terraform module, but the condition was not
adapted. This resulted failures for bare-metal clusters which don't use
Terraform.

This commit places both outputs behind a condition on the Terraform
module to be enabled.

Introduced-by: !1832