fix openstack_enabled for bare-metal
e628268e742fda180448d6696977bdd417b45b1d (!1718) replaced the old
variable on_openstack that was set to true by Terraform templates or had
to be manually set to false in the inventory by the already existing
module option openstack.enabled.

However, it was not considered that in case of openstack.enabled=false,
the inventory file yaook-k8s/group_vars/all/openstack.yaml that exposes
this variable to Ansible, would not be generated at all.
This file should be present in any case so it can be
reliably detected whether openstack is enabled or not.

This commit reworks the erroneous condition, so the file is always
generated.