Add check for openstack.network_mtu
For IPv4 a minimum MTU of 576 is suggested as a reasonable baseline for
the size of datafragments which is why we enforce that [1].
It is unlikely that (cloud) environments with lower MTUs are in use.
We're not even sure if workload (Kubernetes) works with that low MTUs
set.
For IPv6 a minimum MTU of 1280 is technically required [2].

[1] https://datatracker.ietf.org/doc/html/rfc791
[2] https://datatracker.ietf.org/doc/html/rfc8200#section-5

Closes: #781