Ensure ca.key is always restored on CP nodes
If a kubeadm command fails on a control plane node, we always must
ensure the ca.key file is moved back and the kube-controller-manager
manifest is adjusted. Otherwise the control plane node ends up in an
inconsistent state. The "always" section for Ansible blocks is used to
ensure that. For better readability, the respective tasks to hide and
reveal the ca.key have been moved into own task files.