kubeadm: Ignore CreateJob preflight errors
That preflight check triggers a Job to test Kubernetes functionality.
However, it fails when the Job does not complete within 15s. We
frequently observed that the Job takes a bit longer to succeed. The
timeout is not configurable, so we ignore errors of that specific
preflight check such that i.e. Kubernetes upgrades are not aborted due
to the preflight check failing.