Add support for Kubernetes v1.32
... and update mapped components like the snapshot-controller and
Calico. The "CloudDualStackNodeIPs" feature gate went GA [1] which requires
us to implement a migration to remove it from the kubelet configuration
as otherwise kubelet refuses to run. This is only relevant for dual
stack clusters.

To upgrade Calico to the newest version v3.30 it is necessary to
manually apply the CRDs again. It also comes with some new UI called
Whisker [2] which we explicitly disablew. We may add configuration options
for that later.

[1] https://kubernetes.io/docs/reference/command-line-tools-reference/feature-gates-removed/
[2] https://www.tigera.io/blog/calico-whisker-your-new-ally-in-network-observability/