Introduce TAROOK_NIX_FLAGS env var
which allows to pass additional flags to the nix build process of the
inventory. This is a prerequisite for the migration of CI clusters to
another environment as we want to set flavors, images et al
conditionally of the OS_REGION_NAME. To do so, the "--impure" flag is
needed, but it can't be passed to the inventory update script in a clean
way (although there is !1715) as the inventory update script is
triggered in almost every action script.