Support to enable audit logs in Kube-apiserver
We use kubeadm patches for modifying the apiserver's manifest in order
to add some parameters required for enabling audit logging.

The logs are written to the Node's filesystem path
`/var/log/kubernetes/audit/audit.log`.

We also provide a sample policy file. That policy file describes which
api interactions should be logged and which not [1].

Record only events related to pods and api group `apps` (deployment,
daemonset, ...) and also secrets and configmaps.

It's hard defining a proper audit policy since K8s has tons of resources
which might cause negative performance impact due to the high volume
of events.

On the other hand we don't want to miss relevant changes someone or an
entity does to the API. So, we have to find a suitable trade-off.

Therefore we define Google's default, see
https://github.com/kubernetes/kubernetes/blob/cacd595bae429e5739edaf02c6915e9c5731dea7/cluster/gce/gci/configure-helper.sh#L1227C1-L1232C18

Closes #1896

[1]:
https://kubernetes.io/docs/tasks/debug/debug-cluster/audit/#audit-policy

Co-authored-by: Silvio Ankermann <silvio.ankermann@cloudandheat.com>