Support to enable audit logs in Kube-apiserver
We use kubeadm patches for modifying the apiserver's manifest in order
to add some parameters required for enabling audit logging.

The logs are written to the Node's filesystem path `/var/log/kubernetes/audit/audit.log`.

We also provide a sample policy file. That policy file describes which api
interactions should be logged and which not [1].

[1]: https://kubernetes.io/docs/tasks/debug/debug-cluster/audit/#audit-policy

Closes #1896