Audit policy: Apply defaults from Google Container-Optimized OS
It's hard defining a proper audit policy since K8s has tons of resources which
might cause negative performance impact due to the high volume of events.

On the other hand we don't want to miss relevant changes someone or an entity
does to the API. So, we have to find a suitable trade-off.

Therefore we define Google's default, see
https://github.com/kubernetes/kubernetes/blob/cacd595bae429e5739edaf02c6915e9c5731dea7/cluster/gce/gci/configure-helper.sh#L1227C1-L1232C18