Fix nested exceptions in flatten function
The function yk8s-lib.transform.flatten accepts an argument to exempt
certain attributes from flattening. This argument did not work for
nested attributes.

This commit fixes that so that eg. `flatten {except = ["a.b"]}` will do
the right thing. Also, it adds a default value to the argument, so that
it doesn't have to be given.