Conditionally set subnet CIDR variables
As these should only be passed to Terraform and the inventory if the
respecitive IP protocol version is enabled.

This fixes a bug which happens if you have established a Wireguard
tunnel to a dual stack cluster where default values are used for the subnet
CIDRs and you want to connect to a second cluster where e.g. IPv6 is disabled.
The wg-up.sh would complain that a route to the IPv6 network already exists
because the default values are present in the inventory [1] although
IPv6 isn't enabled in that cluster.

[1] https://gitlab.com/yaook/k8s/-/blob/d0d8270d68c96970541f9b1605ce2704e40ac753/actions/wg-up.sh#L21-35