keepalived: Group all VRRP instances including LbaaS_VIPs
... if ch-k8s-lbaas is enabled. This ensures that their states are
always in sync and prevents getting into an inconsistent state. It is
important that the vrrp instance LBaaS_VIPs is in the same state as the
VIP/VIP_v6 (that they are in state "MASTER" on the same node) for traffic
to flow to the LoadBalancer-Service.

A VRRP instance can be part of one group only.