Revert "ci: replace cleanup script with upstream command"
This reverts commit 59e0ee4c36a44d1562b74ed53715d6346c7a6093.

Properly cleaning up an OpenStack project still is a challenge. While
the openstackclient subcommand worked sufficiently well in the previous
CI OpenStack environment, it turned out it does not in the new. It
fails to cleanup all resources but terminates with return code 0
nonetheless. This leads to dangling resources which blow up the quota.
This then breaks CI runs and requires manual cleanup.