bird: Compose filter with functions
With regards to separation of concerns, we want to compose the export
filter for Kubernetes nodes dynamically based upon conditions.
As filters cannot be combined, we have to factor out logic into functions.

Functions get the route table entry as an implicit parameter when called
from within filters which is why it is not stated explicitly.