Refine ch-k8s-lbaas dependencies
As theoretically, ch-k8s-lbaas could be rolled out on bare metal
clusters as well, the dependencies have been refined such that roles
which are only needed when running on OpenStack are not triggered when
runnning not on OpenStack.

Also nftables has been added as ch-k8s-lbaas requires nftables to be set
up when running on OpenStack.