ch-k8s-lbaas: Add option to disable SNAT'ing
In the current setup, LBaaS traffic is routed through the gateway nodes
and these do masquerade the traffic such that the workload behind a
LoadBalancer-Service does not see the real source IP. This commit adds
an option to disable this behavior. For traffic to be able to be
properly returned, it must be routed through the gateways again on the
way back. As Kubernetes nodes are using the OpenStack router as default
gateway, we have to override this such that they are routing through the
gateway by default. This means there is an additional hop! We do use
bird to announce routes to overwrite the default route on Kubernetes
nodes via BGP. We do not directly propagate a route for "0.0.0.0/0" as
that could potentially be ignored. Instead we propagate two routes
"0.0.0.0/1" and "128.0.0.0/1" which has the same effect. It is important
that these routes are not imported into the routing table on the gateway
nodes as that would break connectivity.

This commit is IPv4 specific as we do not support IPv6
LoadBalancer-Services, yet and implementation for that will probably
differ.