Refine audit-logs for kube-apiserver
The previous attempt has been refined such that it is now possible to
finely configure the audit-policy manifest. The "custom_policy" option
has been removed in favor of a generic option which is built properly
with submodules and has defaults. This also ensures the default values
are documented in the reference documentation.

Co-authored-by: Silvio Ankermann <silvio.ankermann@cloudandheat.com>