Switch to kubeadm Configuration v1beta4
which is the default since K8s v1.31. Configurations in an older format
get automatically converted by kubeadm before they get applied, but we
should use v1beta4 right away as that makes it easier to tinker with the
configurations once they are put into a ConfigMap.
The main change which affects us is that the "*.extraArgs" formats have been
reworked.