Don't kubadm init/join initialized nodes
we should never try to init or join a cluster on nodes which already
have been fully initialized Kubernetes wise. This can cause havoc. For
CP nodes, they may try to join the cluster if their local kube-apiserver
is not reachable because it has been restarted recently. The added
condition prevents that.