Fix altering rook/ceph namespace
Some values in the cluster-chart had to be adjusted in order to allow a
user to configure a different namespace than "rook-ceph". Also the
cluster-chart must be aware in which namespace the operator runs such
that it does not try to create resources already created by the
operator.

A differing namespace than the default is now used in the CI.

Closes: #806