Use lists in arguments of mkRenamedOptionModule
Until now, mkRenamedOptionModule accepted section names and option
names in dot notation. While this was easy to type, it made renaming of
options whose names contains dots impossible.

This commit changes the function to accept the path of the option name
as a list. This also allows to integrate the functionality previously
handled by mkRenamedOptionModuleWithNewSection.