renovate: Manage Kubernetes versions
Let the renovate-bot manage the used Kubernetes versions. This includes
the default Kubernetes version configured via nix as well as the
Kubernetes versions used in the CI to test the Kubernetes upgrade path.
As we need manual intervention when adding a new major or minor version
for Kubernetes, only patch version updates are enabled. This requires to
set "separateMinorPatch". This is because minor and patch updates end up
in the same bucket and there is currently no way to have only patch
updates. They always come along with minor updates [1]. The drawback
following from that is, that we will get MRs which suggest to update all
the Kubernetes versions in the CI manifest, which is not useful as we want
to test the whole upgrade path.

  [1]: https://github.com/renovatebot/renovate/discussions/16488