renovate: Manage Kubernetes versions
Let the renovate-bot manage the used Kubernetes versions. This includes
the default Kubernetes version configured via nix as well as the
Kubernetes versions used in the CI to test the Kubernetes upgrade path.
As we need manual intervention when adding a new major or minor version
for Kubernetes, only patch version updates are enabled. This requires to
set "separateMinorPatch".