Move k8s-config to Nix
This commit moves the variables k8s_version_minor and k8s_cri_url to Nix
and removes the k8s-config role.

This is done to prepare the deprecation of the version-map role which
now only contains the version map itself but not the variable
k8s_version_minor that is used in other places, too.