Use lists in arguments of mk*ResourceOptionModule
Until now, mkResourceOptionModule and mkMultiResourceOptionsModule
accepted section names in dot notation and dotless option names. That's
why resource options with dots in their names were impossible to create
and impossible to create in sections whose names contain dots.

This commit changes both functions to accept the path of the section and
option name as a list.