Adjust hotfix preparation to not create changelogs against devel
We do not want to maintain any changelogs on devel. When creating
hotfixes for the latest release, we have to mirror the changes to devel.
However, we want to prevent changelog generation in that case.
This commit ensures that no changelogs are generated.

Related: https://gitlab.com/alasca.cloud/tarook/rdm/-/merge_requests/1