etcd-backup: Drop bucket management tasks
As these are cumbersome to maintain in especially due to the policy of
boto3 that only Amazon S3 object storage is officially supported and
they do not really care about other S3-compatible solutions.

As the user is already required to create the etcd-backup S3
configuration file and put it into Vault, it is feasible to also require
it to create the bucket and configure a lifecycle by that chance.

This makes the local boto3 dependency obsolet. Also this saves us from
ensuring compatibility of the dropped tasks with potentially all S3
solutions out there.

Closes: #818 #815