ch-k8s-lbaas: Move shared secret into Vault
Because if handled via configuration, it ends up world-readable in the
nix-store. Let the orchestrator manage the secret as the orchestrator
configures the agents as well as the controller. For existing clusters,
the configured shared secret is moved into Vault to allow seemless
migration.

If the configuration option is still used, it always overwrites the
value in Vault.

Closes: #499

Co-Authored-By: Bruno Schubert <bruno.schubert@cloudandheat.com>