Check for minimal Kubernetes requirements during bootstrap
As these are enforced by kubeadm. To catch an insufficient control plane
node early, we should check that during bootstrapping.

We cannot really prevent here that Ansible facts are gathered for all
targets hosts if we want to keep the code simple. Targetting the added
play directly against "masters" would not work, as the bootstrap
playbook is already called during gateway preparation. At that point in time,
the masters are not reachable though. To solve that, the added play is
targeted alike the others, and only for nodes in the group "masters" we
do the assertions. For other nodes the task is skipped. However, Ansible
facts are gathered always for all targeted nodes which worsens runtime a
bit. We're basically favoring robustness over runtime here.

Closes: #820