Fix migration of gitignore file
In !2131 (93ea1b4f825c4b9d2666893711780deae23cfa50), we updated the
template for .gitignore to contain markers for the manged parts. But we
missed the fact that the migration script uses the same template and
then adds its own markes.

This commit changes the migration script so that it replaces the whole
section, including the markers, with the template.

It also fixes a bug that prevented the migration script from staging
.gitignore if it had changed (and vice versa).